merges vcx-ci with evernym indy-sdk fork (#15)
* merges vcx-ci with evernym indy-sdk fork


* accepts libindy 1.7.0~942

+++ vcx/ci/scripts/androidBuild.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 setup() {
     echo "Working Directory: ${PWD}"
     set -e
@@ -8,6 +10,7 @@ setup() {
     export PATH=$PATH:/opt/gradle/gradle-3.4.1/bin
     export PATH=${PATH}:$ANDROID_HOME/platform-tools:$ANDROID_HOME/tools:$ANDROID_HOME/build-tools/25.0.2/
     export PATH=${HOME}/.cargo/bin:${PATH}
+    export SOVRIN_REPO=https://repo.sovrin.org/android/libsovtoken/stable
     export VCX_BASE=../vcx
     # For docker
     # export VCX_BASE=${HOME}/vcx
@@ -56,7 +59,7 @@ retrieve_prebuilt_binaries() {
 
 generate_flags(){
     if [ -z $1 ]; then
-        echo "please provide the arch e.g arm, x86 or arm64"
+        echo "please provide the arch e.g arm, arm64, armv7, x86, or x86_64"
         exit 1
     fi
     if [ $1 == "arm" ]; then
@@ -75,14 +78,13 @@ generate_flags(){
         export ARCH="x86"
         export TRIPLET="i686-linux-android"
         export PLATFORM="16"
-        export ABI="x86"
-    fi
-
-    if [ $1 == "arm64" ]; then
-        export ARCH="arm64"
-        export TRIPLET="aarch64-linux-android"
+    elif [ $1 == "x86_64" ]; then
+        export ARCH="x86_64"
+        export TRIPLET="x86_64-linux-android"
         export PLATFORM="21"
-        export ABI="arm64-v8a"
+    else
+        echo "please provide the arch e.g arm, arm64, armv7, x86, or x86_64"
+        exit 1
     fi
 }
 
@@ -91,18 +93,56 @@ get_libindy() {
     if [ -z ${LIBINDY_DIR} ]; then
         [ -z ${LIBINDY_BRANCH} ] && exit 1
         [ -z ${LIBINDY_VERSION} ] && exit 1
-
+		SIMPLE_LIBINDY_VERSION=$(echo ${LIBINDY_VERSION} | cut -f1 -d'-')
         if [ ! -d "libindy_${ARCH}" ]; then
+
             if [ "$LIBINDY_BRANCH" = "stable" ]; then
                 wget https://repo.sovrin.org/android/libindy/${LIBINDY_BRANCH}/${LIBINDY_VERSION}/libindy_android_${ARCH}_${LIBINDY_VERSION}.zip
             else
-                wget https://repo.sovrin.org/android/libindy/${LIBINDY_BRANCH}/${LIBINDY_VERSION}-${LIBINDY_TAG}/libindy_android_${ARCH}_${LIBINDY_VERSION}.zip
+                wget https://repo.sovrin.org/android/libindy/${LIBINDY_BRANCH}/1.7.0-934/libindy_android_${ARCH}_${SIMPLE_LIBINDY_VERSION}.zip
             fi
 
-            unzip libindy_android_${ARCH}_${LIBINDY_VERSION}.zip
+            unzip libindy_android_${ARCH}_${SIMPLE_LIBINDY_VERSION}.zip
 
         fi
         export LIBINDY_DIR="${PWD}/libindy_${ARCH}"
+    fi
+
+}
+
+get_libsovtoken() {
+    set -xv
+    # Todo: This artifact was manually uploaded to this repo. Eventually, the file format will change. That is why it is hardcoded
+    if [ -z ${LIBSOVTOKEN_DIR} ]; then
+        LIBSOVTOKEN_ZIP=libsovtoken_0.9.6-201811211720-4901e95_all.zip
+        if [ ! -d "libsovtoken" ]; then
+            echo "retrieving libsovtoken prebuilt library"
+            wget ${SOVRIN_REPO}/${LIBSOVTOKEN_ZIP}
+            unzip ${LIBSOVTOKEN_ZIP}
+        fi
+        export LIBSOVTOKEN_DIR="${PWD}/libsovtoken/${TRIPLET}"
+    fi
+
+}
+
+get_libnullpay() {
+    set -xv
+    if [ -z ${LIBNULLPAY_DIR} ]; then
+        [ -z ${LIBINDY_BRANCH} ] && exit 1
+        [ -z ${LIBINDY_VERSION} ] && exit 1
+		SIMPLE_LIBINDY_VERSION=$(echo ${LIBINDY_VERSION} | cut -f1 -d'-')
+        if [ ! -d "libnullpay_${ARCH}" ]; then
+
+            if [ "$LIBINDY_BRANCH" = "stable" ]; then
+                wget https://repo.sovrin.org/android/libnullpay/${LIBINDY_BRANCH}/${LIBINDY_VERSION}/libnullpay_android_${ARCH}_${LIBINDY_VERSION}.zip
+            else
+                wget https://repo.sovrin.org/android/libnullpay/${LIBINDY_BRANCH}/1.7.0-934/libnullpay_android_${ARCH}_${SIMPLE_LIBINDY_VERSION}.zip
+            fi
+
+            unzip libnullpay_android_${ARCH}_${SIMPLE_LIBINDY_VERSION}.zip
+
+        fi
+        export LIBNULLPAY_DIR="${PWD}/libnullpay_${ARCH}"
     fi
 
 }
@@ -119,10 +159,18 @@ build_vcx() {
         echo "missing libindy_${ARCH} directory. Cannot proceed without it."
         exit 1
     fi
+    if [ ! -d ${LIBSOVTOKEN_DIR} ]; then
+        echo "missing libsovtoken directory. Cannot proceed without it."
+        exit 1
+    fi
+    if [ ! -d ${LIBNULLPAY_DIR} ]; then
+        echo "missing libnullpay directory. Cannot proceed without it."
+        exit 1
+    fi
 
     pushd ${LIBVCX_PATH}
     mkdir -p toolchains/
-    ./build.nondocker.sh ${ARCH} ${PLATFORM} ${TRIPLET} ${OPENSSL_DIR} ${SODIUM_DIR} ${LIBZMQ_DIR} ${LIBINDY_DIR}
+    ./build.nondocker.sh ${ARCH} ${PLATFORM} ${TRIPLET} ${OPENSSL_DIR} ${SODIUM_DIR} ${LIBZMQ_DIR} ${LIBINDY_DIR} ${LIBSOVTOKEN_DIR} ${LIBNULLPAY_DIR}
     popd
     rm -rf libvcx_${ARCH}
     mv ${LIBVCX_PATH}libvcx_${ARCH} .
@@ -131,4 +179,6 @@ build_vcx() {
 
 setup $1
 get_libindy $1
+get_libsovtoken
+get_libnullpay
 build_vcx $1
